Cleanup sample and refactor to only have a single kotlin library, but both a Maven and Gradle App module as well.
