+++ .gitignore
@@ -1,4 +1,3 @@
-
 # Created by https://www.gitignore.io/api/c,git,c++,java,cmake,xcode,kotlin,android,intellij,visualstudiocode
 # Edit at https://www.gitignore.io/?templates=c,git,c++,java,cmake,xcode,kotlin,android,intellij,visualstudiocode
 
@@ -327,3 +326,6 @@ DerivedData/
 **/xcshareddata/WorkspaceSettings.xcsettings
 
 # End of https://www.gitignore.io/api/c,git,c++,java,cmake,xcode,kotlin,android,intellij,visualstudiocode
+
+# Custom 
+JavaKotlinMavenInterop/java-app-maven/target
