--- JavaInterop/java-lib/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-    id 'java'
-    id 'application'
-}
-
-group = 'org.example'
-version = '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation(project(":kotlin-lib"))
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-application {
-    mainClass = "org.example.Main"
-}
-
-test {
-    useJUnitPlatform()
-}
