--- JavaInterop/java-lib/src/main/java/org/example/Main.java
@@ -1,8 +1,0 @@
-package org.example;
-
-public class Main {
-    public static void main(String[] args) throws InterruptedException {
-        RealmTestRunner runner = new RealmTestRunner();
-        runner.start();
-    }
-}
