--- JavaInterop/settings.gradle
@@ -1,4 +1,0 @@
-rootProject.name = 'JavaInterop'
-include 'java-lib'
-include 'kotlin-lib'
-
