+++ JavaKotlinMavenInterop/settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = 'JavaKotlinMavenInterop'
+include 'kotlin-lib'
+include 'java-app-gradle'
+
